Contrastive Cross Covariance/Correlation Analysis

This function performs contrastive cross covariance/correlation analysis (3CA) or sparse contrastive cross covariance/correlation (s3CA), in comparison, also perform target/ancillary group only (sparse) cross covariance analysis.

cCCA(
  Xa,
  Ya,
  Xt,
  Yt,
  eta = NULL,
  sparse_ctst = NULL,
  sparse_trt = NULL,
  sparse_ctrl = NULL
)

Arguments

Xa

data frame for one data type in ancillary group, samples in rows, variables in columns. To conduct analysis based on cross corrlation matrix, each feature should be centered and standardized.

Ya

data frame for another data type in ancillary group, samples in rows, variables in columns. To conduct analysis based on cross corrlation matrix, each feature should be centered and standardized.

Xt

data frame for one data type in target group, samples in rows, variables in columns. To conduct analysis based on cross corrlation matrix, each feature should be centered and standardized.

Yt

data frame for another data type in target group, samples in rows, variables in columns. To conduct analysis based on cross corrlation matrix, each feature should be centered and standardized.

eta

tuning parameter controlling how much ancillary variation should be contrasted off from the target variation. If NULL, the estimated eta introduced in paper will be used.

sparse_ctst

parameter controlling the sparsity of s3CA, if NULL, use 3CA, else use s3CA. For use of s3CA, if vector, refer sumabss in PMD.cv from package PMA, if single number, refer sumabs in PMD in package PMA.

sparse_trt

parameter controlling the sparsity of target sparse cross covariance/correlation analysis, if NULL, use cross covariance/correlation analysis, else use sparse cross covariance/correlation analysis. For use of sparse version, if vector, refer sumabss in PMD.cv from package PMA, if single number, refer sumabs in PMD in package PMA.

sparse_ctrl

parameter controlling the sparsity of ancillary sparse cross covariance/correlation analysis, if NULL, use cross covariance/correlation analysis, else use sparse cross covariance/correlation analysis. For use of sparse version, if vector, refer sumabss in PMD.cv from package PMA, if single number, refer sumabs in PMD in package PMA.

Value

A list containing contrastive SVD object CCA_ctst including singular values d, loadings for X u, loadings for Y v; target only SVD object CCA_trt and ancillary only SVD object CCA_ctrl .

Examples

library(ConSpec)
data(intro)
res_3CA <- cCCA(Xa = intro$cCCA$Xa, Ya = intro$cCCA$Ya,
Xt = intro$cCCA$Xt, Yt = intro$cCCA$Yt, eta  = NULL,
sparse_ctst = NULL, sparse_trt = NULL, sparse_ctrl = NULL)
#> 
#>  optimal eta is 2.36543920954191