Simulation function for contrastive spectral clustering (CSC)

This function performs simulation to prove the interesting of hidden clusters by using cPCA and eta tuning algorithm. Same as in the example of Section 2, we consider two target clusters and two nuisance clusters

sim_CSC(
  n11,
  n12,
  n21,
  n22,
  m1tilde,
  m2tilde,
  mu1,
  mu2,
  theta1,
  theta2,
  p1,
  p2,
  sigma_t = 1,
  sigma_a = 1,
  Sigma_t = NULL,
  Sigma_a = NULL,
  eta,
  repetition = 10,
  save_plot = FALSE
)

Arguments

n11

number of target samples in interesting cluster 1 and nuisance cluster 1.

n12

number of target samples in interesting cluster 1 and nuisance cluster 2.

n21

number of target samples in interesting cluster 2 and nuisance cluster 1.

n22

number of target samples in interesting cluster 2 and nuisance cluster 2.

m1tilde

number of ancillary samples in nuisance cluster 1.

m2tilde

number of ancillary samples in nuisance cluster 2.

mu1

non-zero values in mu_t1.

mu2

non-zero values in mu_t2.

theta1

non-zero values in theta_1.

theta2

non-zero values in theta_2

p1

number of non-zero elements in mu_t1 and mu_t2.

p2

number of non-zero elements in theta_1 and theta_2.

sigma_t

standard deviation of the uncorrelated errors in target group.

sigma_a

standard deviation of the uncorrelated errors in ancillary group.

Sigma_t

the error covariance matrix in target group, if is NULL, use standard deviation specified and assume uncorrelation.

Sigma_a

the error covariance matrix in ancillary group, if is NULL, use standard deviation specified and assume uncorrelation.

eta

if eta is a vector, eta_tuning algorithm will be called to select optimal eta; if a number, eta will be used.

repetition

number of replications of the simulation.

save_plot

logical, whether save the score plot and loading plot for the first replication.

Value

A data frame containing the selected optimal eta, the four metrics values.

Examples

# See vignettes/ContrastiveSpectralClustering.Rmd for example usage.